[feat] attraction 상세보기(findById)

+++ bonvoyageW/src/main/java/com/bonvoyage/domain/attraction/dto/AttractionInfoDto.java
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class AttractionInfoDto {
 
-    private int contentId;
+    private long contentId;
     private String title;
     private String addr1;
     private String tel;
