+++ bonvoyageW/src/main/java/com/bonvoyage/domain/attraction/repository/AttractionInfoRepository.java
@@ -5,6 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AttractionInfoRepository extends JpaRepository<AttractionInfoEntity, Integer> {
-//    List<AttractionInfoEntity> findAll();
+public interface AttractionInfoRepository extends JpaRepository<AttractionInfoEntity, Long> {
 }
